fea: 권장 영양소 조회 구현

+++ backend/Foody/src/main/java/com/foody/recommendednutrient/entity/RecommendedNutrient.java
@@ -13,14 +13,14 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class RecommendedNutrient extends BaseEntity {
 
-    private float energy; // 칼로리
-    private float carbohydrates; // 탄수화물
-    private float protein; // 단백질
-    private float dietaryFiber; // 식이섬유
-    private float calcium; // 칼슘
-    private float sodium; // 나트륨
-    private float iron; // 철분
-    private float fats; // 지방
-    private float vitaminA; // 비타민A
-    private float vitaminC; // 비타민C
+    private double energy; // 칼로리
+    private double carbohydrates; // 탄수화물
+    private double protein; // 단백질
+    private double dietaryFiber; // 식이섬유
+    private double calcium; // 칼슘
+    private double sodium; // 나트륨
+    private double iron; // 철분
+    private double fats; // 지방
+    private double vitaminA; // 비타민A
+    private double vitaminC; // 비타민C
 }
